+++ spec/examples/skip_spec.rb
@@ -1,0 +1,14 @@
+# frozen_string_literal: true
+
+require './lib/rast'
+
+class Skip
+end
+
+xrast Skip do
+end
+
+rast String do
+  xspec '#skip_rspec' do
+  end
+end